+++ src/main/groovy/com/ullink/AssemblyInfoVersionPatcher.groovy
@@ -6,7 +6,6 @@ import org.gradle.api.provider.ListProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.InputFiles
-import org.gradle.api.tasks.OutputFiles
 import org.gradle.api.tasks.TaskAction
 
 class AssemblyInfoVersionPatcher extends DefaultTask {