+++ src/main/groovy/com/ullink/IExecutableResolver.groovy
@@ -3,5 +3,7 @@ package com.ullink
 interface IExecutableResolver {
     ProcessBuilder executeDotNet(File exe)
 
+    ProcessBuilder executeDotNetApp(File exe)
+
     void setupExecutable(Msbuild msbuild);
 }